% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthrograph function.R
\name{anthrograph}
\alias{anthrograph}
\title{Upload to database}
\usage{
anthrograph(finaltable)
}
\description{
This function allow you to create an anthropometric clinical graph.
}
\keyword{anthrograph}

