% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANTHROPOMETRICS_SOURCE_CDC.R
\docType{data}
\name{ANTHROPOMETRICS_SOURCE_CDC}
\alias{ANTHROPOMETRICS_SOURCE_CDC}
\title{Anthropometrics Source CDC}
\format{Source data for calculating z-scores and percentiles}
\usage{
ANTHROPOMETRICS_SOURCE_CDC
}
\description{
Anthropometrics Source
}
\keyword{CDC}
\keyword{anthropometrics}
\keyword{source}

