% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANTHROPOMETRICS_SOURCE_NHANES.R
\docType{data}
\name{ANTHROPOMETRICS_SOURCE_NHANES}
\alias{ANTHROPOMETRICS_SOURCE_NHANES}
\title{Anthropometrics Source NHANES}
\format{Source data for calculating z-scores and percentiles}
\usage{
ANTHROPOMETRICS_SOURCE_NHANES
}
\description{
Anthropometrics Source
}
\keyword{NHANES}
\keyword{anthropometrics}
\keyword{source}

