% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANTHROPOMETRICS_SOURCE_WHO.R
\docType{data}
\name{ANTHROPOMETRICS_SOURCE_WHO}
\alias{ANTHROPOMETRICS_SOURCE_WHO}
\title{Anthropometrics Source WHO}
\format{Source data for calculating z-scores and percentiles}
\usage{
ANTHROPOMETRICS_SOURCE_WHO
}
\description{
Anthropometrics Source
}
\keyword{WHO}
\keyword{anthropometrics}
\keyword{source}

