% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_anthro function.R
\name{calculate_anthro}
\alias{calculate_anthro}
\title{Calculate Anthropometrics}
\usage{
calculate_anthro()
}
\description{
This function allows you to calculate anthropometric calculations, z-scores, percentiles, and it creates a clinical anthropometric graph. It also gives you the option to create an excel file to check the values that are used to calculate the z-scores and percentiles, and the option to save another excel file to check for any data flags.
}
\examples{
calculate_anthro()
}
\keyword{anthropometrics}

