% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadanthros.R
\name{uploadanthros}
\alias{uploadanthros}
\title{Upload to database}
\usage{
uploadanthros(finaltable)
}
\arguments{
\item{data}{Data to be input into the research table}

\item{str}{String defined as either "MED" or "SEIZURE"}

\item{sourcedata}{Data to be input to the calculated table}

\item{table}{Data to be added to the ranking table (only used if str=="SEIZURE")}
}
\description{
This function allow you to upload seizure or med data into the clinical and research tables of the MySQL database.
It will be necessary to have access to the lab's database in order to do this
}
\keyword{database}
\keyword{upload}

